refactor : splitter 패키지 이동
