+++ src/main/java/calculator/splitter/CustomDelimiterSplitter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.splitter;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
