+++ src/main/java/calculator/splitter/DefaultDelimiterSplitter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.splitter;
 
 public class DefaultDelimiterSplitter extends AbstractDelimiterSplitter {
