+++ src/main/java/calculator/splitter/DelegateDelimiterSplitter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.splitter;
 
 import java.util.ArrayList;
 import java.util.Collections;
